+++ wallet-database/src/repositories/bill.rs
@@ -103,7 +103,7 @@ impl BillRepo {
         min_value: Option<f64>,
         start: Option<i64>,
         end: Option<i64>,
-        transfer_type: Vec<i64>,
+        transfer_type: Vec<i32>,
         page: i64,
         page_size: i64,
         pool: &DbPool,
